+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `Projectionist` to play, boot and retire projectors
+- `ProjectorStateLedger` that tracks projector state information
+- `Projection`, `Projector` and `EventHandlingStrategy` interfaces
 
 ## [v0.2.0] 2018-08-21
 ### Added